+++ src/lib.rs
@@ -1,4 +1,4 @@
 pub mod gen;
 pub mod sem;
 pub mod syntax;
-pub mod typing;
+pub mod ty;
